+++ annotations/src/main/java/react4j/annotations/Prop.java
@@ -1,0 +1,35 @@
+package react4j.annotations;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Target;
+import javax.annotation.Nonnull;
+
+/**
+ * Annotation used to specify an abstract method that returns a prop.
+ * The property is extracted from Reacts underlying props object.
+ *
+ * <p>The method that is annotated with <code>@Prop</code> must also comply with the following constraints:</p>
+ * <ul>
+ * <li>Must not be annotated with any other react annotation</li>
+ * <li>Must have 0 parameters</li>
+ * <li>Must return a value</li>
+ * <li>Must be an abstract instance method</li>
+ * <li>Must not throw exceptions</li>
+ * </ul>
+ */
+@Documented
+@Target( ElementType.METHOD )
+public @interface Prop
+{
+  /**
+   * Return the name of the prop.
+   * If the underlying method conforms to java beans accessor conventions (i.e. starts with "is" and is a boolean
+   * or starts with "get") then the name is the same as the java bean convention dictates, otherwise the name of
+   * the method is used as the default value for the prop.
+   *
+   * @return the name of the prop.
+   */
+  @Nonnull
+  String name() default "<default>";
+}