+++ processor/src/main/java/react4j/processor/Constants.java
@@ -8,6 +8,7 @@ final class Constants
   static final String JS_FUNCTION_CLASSNAME = "jsinterop.annotations.JsFunction";
   static final String COMPONENT_CLASSNAME = "react4j.core.Component";
   static final String EVENT_HANDLER_ANNOTATION_CLASSNAME = "react4j.annotations.EventHandler";
+  static final String PROP_ANNOTATION_CLASSNAME = "react4j.annotations.Prop";
   static final String NO_AUTO_ACTION_ANNOTATION_CLASSNAME = "react4j.arez.NoAutoAction";
   static final String REACT_COMPONENT_ANNOTATION_CLASSNAME = "react4j.annotations.ReactComponent";
   static final String INJECT_ANNOTATION_CLASSNAME = "javax.inject.Inject";