+++ processor/src/test/resources/expected/com/example/prop/PropTypeInt_.java
@@ -1,0 +1,59 @@
+package com.example.prop;
+
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import jsinterop.base.Js;
+import react4j.core.BaseContext;
+import react4j.core.BaseProps;
+import react4j.core.BaseState;
+import react4j.core.ComponentConstructorFunction;
+import react4j.core.NativeAdapterComponent;
+import react4j.core.React;
+import react4j.core.ReactConfig;
+import react4j.core.ReactNode;
+
+@Generated("react4j.processor.ReactProcessor")
+class PropTypeInt_ extends PropTypeInt {
+  private static final ComponentConstructorFunction<BaseProps, BaseContext> TYPE = getConstructorFunction();
+
+  @Nonnull
+  static ReactNode _create() {
+    return React.createElement( TYPE );
+  }
+
+  @Nonnull
+  static ReactNode _create(@Nullable final BaseProps props) {
+    return React.createElement( TYPE, props );
+  }
+
+  @Nonnull
+  static ReactNode _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
+    return React.createElement( TYPE, props, child );
+  }
+
+  @Nonnull
+  private static ComponentConstructorFunction<BaseProps, BaseContext> getConstructorFunction() {
+    final ComponentConstructorFunction<BaseProps, BaseContext> componentConstructor = NativeReactComponent::new;
+    if ( ReactConfig.enableComponentNames() ) {
+      Js.asPropertyMap( componentConstructor ).set( "displayName", "PropTypeInt" );
+    }
+    return componentConstructor;
+  }
+
+  @Override
+  protected int getMyProp() {
+    return Js.asPropertyMap( props() ).getAny( "myProp" ).asInt();
+  }
+
+  private static final class NativeReactComponent extends NativeAdapterComponent<BaseProps, BaseState, BaseContext, PropTypeInt> {
+    NativeReactComponent(@Nullable final BaseProps props, @Nullable final BaseContext context) {
+      super( props, context );
+    }
+
+    @Override
+    protected PropTypeInt createComponent() {
+      return new PropTypeInt_();
+    }
+  }
+}